t_elapsed is now double; GraphicText can be set from XML details
t_elapsed is now double overall in all Update() methods.
GraphicText's contents can be set with the attribute content in an xml
file.

+++ jni/ai/mjAutomaton.cpp
@@ -17,7 +17,7 @@ mjAutomaton::~mjAutomaton()
     //dtor..
 }
 
-void mjAutomaton::Update(float t_elapsed)
+void mjAutomaton::Update(double t_elapsed)
 {
     if (currentState)
     {
