+++ jni/game/entities/BatBot.h
@@ -54,11 +54,11 @@ class BatBot: public KosmoObject, public mjAutomaton
 
         bool ignoresGravity = true;
 
-        virtual void Update(float t_elapsed) override;
+        virtual void Update(double t_elapsed) override;
 
         //virtual void ProcessCollisionEffects() override;
 
-        virtual void ProcessPhysicsEffects(float t_elapsed) override;
+        virtual void ProcessPhysicsEffects(double t_elapsed) override;
         virtual void SetDetailsFromXML(XMLElement* entity) override;
