+++ jni/graphics/mjSceneGraph.h
@@ -42,7 +42,7 @@ public:
     void Initialize(mjResourceManager* resourceManager);
     void Add(mjObject* object, bool isDrawable=true, bool castsShadow=false, bool isTranslucent=false);
     void Reclassify(mjObject* object);
-    void Update(float t_elapsed);
+    void Update(double t_elapsed);
     //void AddGroup(std::vector<mjObject*>* group, bool toDrawable=true, bool toShadowCasters=false, bool toTranslucent=false);
     //void RemoveGroup(std::vector<mjObject *>* group);
